Booking system for the massage.
1.about.php
The page describes about the oraganization
2.Booking_Canceling.php
To cancel an already existing Appointment.
3.config.php
The database configuration of the front-end and back-end. This intializes the database sample.
4.contact.php
Displays the details of the manager and the booking officer.
5.header.php
This has the code of Bootstrap and the style format for the navbar,etc...
6.Home_c.php
This displays the user details and his previous booking and the new booking.

7.index.php
The Main Homepage of the website and it has the sign in form and all the details of our website as the links in the side bar and nav bar.
8.login_verify.php
The page is used for the user varification with the log-In table in the database. query to check the password and username.
9.logout.php
It is used for the Signout function fro the individual users
10.page1.php
Explains the user what is sports massage
11.page2.php
Describes the user what is hot stone massage
12.page3.php
Gives the user the idea about the Cupping therapy
13.page4.php
Tells the user what is kinesio therapy and its benefits
14.Signup.php
Registration Form where the user types the information
15.Signup_insert.php
Inserts the user registration into the database
16.Booking_Cancel.php
For any appointment cancellation
17.storedata.php
it stores the data into the database.And the booking date, time and type massage.